# specs/001-school-library-system/contracts/api-spec.yaml
openapi: 3.0.3
info:
  title: BCD Library Management System API
  description: |
    REST API for school library management (Bibliothèque Centre Documentaire).

    **Operational Model**: Librarian-operated system with offline-first design.
    - No fines/fees (privilege blocking instead)
    - No email/SMS notifications (printed reports)
    - Librarian-mediated access (no student self-service)

    **Key Features**:
    - Circulation (checkout/return/renew with barcode scanning)
    - Cataloging (BNF SRU API integration for ISBN lookup)
    - Borrower management (students/teachers/staff organized by class)
    - Holds/reservations (librarian-mediated queue management)
    - Reports (overdue, statistics, never-borrowed)
    - Barcode generation (Code 39/128)
  version: 1.0.0
  contact:
    name: BCD API Support
    email: support@bcd-library.local
  license:
    name: MIT
servers:
  - url: http://localhost:8000/api/v1
    description: Development server (localhost)
  - url: http://library-server.local:8000/api/v1
    description: Production server (local network)
...
paths:
...
  /holds/borrower/{borrower_id}:
    get:
      tags: [holds]
      summary: Get borrower's holds
      description: Returns all active holds for a borrower
      operationId: getBorrowerHolds
      parameters:
        - name: borrower_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Borrower's active holds
          content:
            application/json:
              schema:
                type: object
                properties:
                  borrower:
                    $ref: '#/components/schemas/BorrowerSummary'
                  holds:
                    type: array
                    items:
                      type: object
                      properties:
                        hold_id:
                          type: integer
                        bibliographic_record_id:
                          type: integer
                        title:
                          type: string
                        queue_position:
                          type: integer
                        status:
                          type: string
                        hold_date:
                          type: string
                          format: date-time
                        estimated_availability:
                          type: string
                          format: date
                          nullable: true
        '404':
          $ref: '#/components/responses/NotFound'
...
  /borrowers:
    get:
      tags: [borrowers]
      summary: List borrowers
      operationId: listBorrowers
      parameters:
        - name: class_id
          in: query
          schema:
            type: integer
        - name: role
          in: query
          schema:
            type: string
            enum: [student, teacher, staff]
        - name: active
          in: query
          schema:
            type: boolean
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            maximum: 100
      responses:
        '200':
          description: List of borrowers
          content:
            application/json:
              schema:
                type: object
                properties:
                  total:
                    type: integer
                  results:
                    type: array
                    items:
                      $ref: '#/components/schemas/Borrower'

    post:
      tags: [borrowers]
      summary: Create borrower
      operationId: createBorrower
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BorrowerCreate'
      responses:
        '201':
          description: Borrower created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Borrower'
        '400':
          $ref: '#/components/responses/BadRequest'
  /borrowers/{borrower_id}:
    get:
      tags: [borrowers]
      summary: Get borrower details
      operationId: getBorrower
      parameters:
        - name: borrower_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Borrower details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BorrowerDetailed'
        '404':
          $ref: '#/components/responses/NotFound'
  /borrowers/import:
    post:
      tags: [borrowers]
      summary: Import borrowers from CSV
      operationId: importBorrowers
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Import results
          content:
            application/json:
              schema:
                type: object
                properties:
                  imported:
                    type: integer
                  skipped:
                    type: integer
                  errors:
                    type: array
                    items:
                      type: object

  # ========================================
  # REPORT ENDPOINTS
  # ========================================
...
components:
  schemas:
...
    Borrower:
      type: object
      properties:
        id:
          type: integer
        borrower_id:
          type: string
        first_name:
          type: string
        last_name:
          type: string
        full_name:
          type: string
        role:
          type: string
          enum: [student, teacher, staff]
        class_id:
          type: integer
          nullable: true
        grade_level:
          type: string
          nullable: true
        barcode:
          type: string
        active:
          type: boolean
        blocked_reason:
          type: string
          nullable: true
    BorrowerCreate:
      type: object
      required:
        - borrower_id
        - first_name
        - last_name
        - role
      properties:
        borrower_id:
          type: string
        first_name:
          type: string
        last_name:
          type: string
        role:
          type: string
          enum: [student, teacher, staff]
        class_id:
          type: integer
          nullable: true
        email:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
    BorrowerSummary:
      type: object
      properties:
        borrower_id:
          type: string
        full_name:
          type: string
        class_name:
          type: string
          nullable: true
        current_loans:
          type: integer
        loan_limit:
          type: integer
        active:
          type: boolean
        blocked_reason:
          type: string
          nullable: true
    BorrowerDetailed:
      allOf:
        - $ref: '#/components/schemas/Borrower'
        - type: object
          properties:
            current_loans:
              type: array
              items:
                type: object
            active_holds:
              type: array
              items:
                type: object
            circulation_stats:
              type: object
              properties:
                total_checkouts:
                  type: integer
                total_overdue:
                  type: integer
...
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
          nullable: true
  responses:
    BadRequest:
      description: Bad request (validation error)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "validation_error"
            message: "Invalid request parameters"
            details:
              field: "borrower_id"
              issue: "Required field missing"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "not_found"
            message: "Resource not found"